Boilerplate Workflow

This is a boilerplate function to create automatically the following:

recipe

model specification

workflow

tuned model (grid ect)

calibration tibble and plot

Usage

ts_auto_glmnet(
  .data,
  .date_col,
  .value_col,
  .formula,
  .rsamp_obj,
  .prefix = "ts_glmnet",
  .tune = TRUE,
  .grid_size = 10,
  .num_cores = 1,
  .cv_assess = 12,
  .cv_skip = 3,
  .cv_slice_limit = 6,
  .best_metric = "rmse",
  .bootstrap_final = FALSE
)

Arguments

.data

The data being passed to the function. The time-series object.

.date_col

The column that holds the datetime.

.value_col

The column that has the value

.formula

The formula that is passed to the recipe like value ~ .

.rsamp_obj

The rsample splits object

.prefix

Default is ts_glmnet

.tune

Defaults to TRUE, this creates a tuning grid and tuned model.

.grid_size

If .tune is TRUE then the .grid_size is the size of the tuning grid.

.num_cores

How many cores do you want to use. Default is 1

.cv_assess

How many observations for assess. See timetk::time_series_cv()

.cv_skip

How many observations to skip. See timetk::time_series_cv()

.cv_slice_limit

How many slices to return. See timetk::time_series_cv()

.best_metric

Default is "rmse". See modeltime::default_forecast_accuracy_metric_set()

.bootstrap_final

Not yet implemented.

Value

A list

Details

This uses parsnip::linear_reg() and sets the engine to glmnet

See also

Author

Steven P. Sanderson II, MPH

Examples

# \donttest{
library(dplyr)
library(timetk)
library(modeltime)

data <- AirPassengers %>%
  ts_to_tbl() %>%
  select(-index)

splits <- time_series_split(
  data
  , date_col
  , assess = 12
  , skip = 3
  , cumulative = TRUE
)

ts_glmnet <- ts_auto_glmnet(
  .data = data,
  .num_cores = 2,
  .date_col = date_col,
  .value_col = value,
  .rsamp_obj = splits,
  .formula = value ~ .,
  .grid_size = 5
)
#> `geom_smooth()` using method = 'loess' and formula = 'y ~ x'
#> Warning: span too small.   fewer data values than degrees of freedom.
#> Warning: pseudoinverse used at -5.2237
#> Warning: neighborhood radius 0.80263
#> Warning: reciprocal condition number  0
#> Warning: There are other near singularities as well. 3.442
#> Warning: span too small.   fewer data values than degrees of freedom.
#> Warning: pseudoinverse used at 0.04625
#> Warning: neighborhood radius 0.15375
#> Warning: reciprocal condition number  0
#> Warning: There are other near singularities as well. 0.36451
#> Warning: span too small.   fewer data values than degrees of freedom.
#> Warning: pseudoinverse used at -5.2237
#> Warning: neighborhood radius 0.80263
#> Warning: reciprocal condition number  0
#> Warning: There are other near singularities as well. 3.442
#> Warning: span too small.   fewer data values than degrees of freedom.
#> Warning: pseudoinverse used at 0.04625
#> Warning: neighborhood radius 0.15375
#> Warning: reciprocal condition number  0
#> Warning: There are other near singularities as well. 0.36451
#> Warning: span too small.   fewer data values than degrees of freedom.
#> Warning: pseudoinverse used at -5.2237
#> Warning: neighborhood radius 0.80263
#> Warning: reciprocal condition number  0
#> Warning: There are other near singularities as well. 3.442
#> Warning: span too small.   fewer data values than degrees of freedom.
#> Warning: pseudoinverse used at 0.04625
#> Warning: neighborhood radius 0.15375
#> Warning: reciprocal condition number  0
#> Warning: There are other near singularities as well. 0.36451
#> Warning: span too small.   fewer data values than degrees of freedom.
#> Warning: pseudoinverse used at -5.2237
#> Warning: neighborhood radius 0.80263
#> Warning: reciprocal condition number  0
#> Warning: There are other near singularities as well. 3.442
#> Warning: span too small.   fewer data values than degrees of freedom.
#> Warning: pseudoinverse used at 0.04625
#> Warning: neighborhood radius 0.15375
#> Warning: reciprocal condition number  0
#> Warning: There are other near singularities as well. 0.36451
#> Warning: span too small.   fewer data values than degrees of freedom.
#> Warning: pseudoinverse used at -5.2237
#> Warning: neighborhood radius 0.80263
#> Warning: reciprocal condition number  0
#> Warning: There are other near singularities as well. 3.442
#> Warning: span too small.   fewer data values than degrees of freedom.
#> Warning: pseudoinverse used at 0.04625
#> Warning: neighborhood radius 0.15375
#> Warning: reciprocal condition number  0
#> Warning: There are other near singularities as well. 0.36451
#> Warning: span too small.   fewer data values than degrees of freedom.
#> Warning: pseudoinverse used at -5.2237
#> Warning: neighborhood radius 0.80263
#> Warning: reciprocal condition number  0
#> Warning: There are other near singularities as well. 3.442
#> Warning: span too small.   fewer data values than degrees of freedom.
#> Warning: pseudoinverse used at 0.04625
#> Warning: neighborhood radius 0.15375
#> Warning: reciprocal condition number  0
#> Warning: There are other near singularities as well. 0.36451

ts_glmnet$recipe_info
#> $recipe_call
#> recipe(.data = data, .date_col = date_col, .value_col = value, 
#>     .formula = value ~ ., .rsamp_obj = splits, .grid_size = 5, 
#>     .num_cores = 2)
#> 
#> $recipe_syntax
#> [1] "ts_glmnet_recipe <-"                                                                                                                                
#> [2] "\n  recipe(.data = data, .date_col = date_col, .value_col = value, .formula = value ~ \n    ., .rsamp_obj = splits, .grid_size = 5, .num_cores = 2)"
#> 
#> $rec_obj
#> 
#> ── Recipe ──────────────────────────────────────────────────────────────────────
#> 
#> ── Inputs 
#> Number of variables by role
#> outcome:   1
#> predictor: 1
#> 
#> ── Operations 
#> • Timeseries signature features from: date_col
#> • Holiday signature features from: date_col
#> • Novel factor level assignment for: recipes::all_nominal_predictors()
#> • Variable mutation for: tidyselect::vars_select_helpers$where(is.character)
#> • Variable mutation for: as.numeric(^date_col)
#> • Dummy variables from: recipes::all_nominal()
#> • Zero variance filter on: recipes::all_predictors(), -date_col_index.num
#> • Centering and scaling for: recipes::all_numeric_predictors(), ...
#> 
# }