Time Series Model Spec Template

This function will create a generic tuneable model specification, this function can be used by itself and is called internally by ts_model_auto_tune().

Usage

ts_model_spec_tune_template(.parsnip_engine = NULL, .model_spec_class = NULL)

Arguments

.parsnip_engine

The model engine that is used by parsnip::set_engine().

.model_spec_class

The model spec class that is use by parsnip. For example the 'kernlab' engine can use both svm_poly and svm_rbf.

Value

A tuneable parsnip model specification.

Details

This function takes in a single parameter and uses that to output a generic tuneable model specification. This function can work with the following parsnip/modeltime engines:

"auto_arima"

"auto_arima_xgboost"

"ets"

"croston"

"theta"

"smooth_es"

"stlm_ets"

"tbats"

"stlm_arima"

"nnetar"

"prophet"

"prophet_xgboost"

"lm"

"glmnet"

"stan"

"spark"

"keras"

"earth"

"xgboost"

"kernlab"

See also

Author

Steven P. Sanderson II, MPH

Examples

ts_model_spec_tune_template("ets")
#> Exponential Smoothing State Space Model Specification (regression)
#> 
#> Main Arguments:
#>   seasonal_period = auto
#>   error = auto
#>   trend = auto
#>   season = auto
#>   damping = auto
#>   smooth_level = tune::tune()
#>   smooth_trend = tune::tune()
#>   smooth_seasonal = tune::tune()
#> 
#> Computational engine: ets 
#> 
ts_model_spec_tune_template("prophet")
#> PROPHET Regression Model Specification (regression)
#> 
#> Main Arguments:
#>   changepoint_num = tune::tune()
#>   changepoint_range = tune::tune()
#>   seasonality_yearly = auto
#>   seasonality_weekly = auto
#>   seasonality_daily = auto
#>   prior_scale_changepoints = tune::tune()
#>   prior_scale_seasonality = tune::tune()
#>   prior_scale_holidays = tune::tune()
#> 
#> Computational engine: prophet 
#>