Random Walk Function

This function takes in four arguments and retunrs a tibble of random walks.

ts_random_walk(
  .mean = 0,
  .sd = 0.1,
  .num_walks = 100,
  .periods = 100,
  .initial_value = 1000
)

Arguments

.mean
The desired mean of the random walks

.sd
The standard deviation of the random walks

.num_walks
The number of random walks you want generated

.periods
The length of the random walk(s) you want generated

.initial_value
The intial value where the random walks should start

Value

A tibble

Details

Monte Carlo simulations were first formally designed in the 1940’s while developing nuclear weapons, and since have been heavily used in various fields to use randomness solve problems that are potentially deterministic in nature. In finance, Monte Carlo simulations can be a useful tool to give a sense of how assets with certain characteristics might behave in the future. While there are more complex and sophisticated financial forecasting methods such as ARIMA (Auto-Regressive Integrated Moving Average) and GARCH (Generalised Auto-Regressive Conditional Heteroskedasticity) which attempt to model not only the randomness but underlying macro factors such as seasonality and volatility clustering, Monte Carlo random walks work surprisingly well in illustrating market volatility as long as the results are not taken too seriously.

Author

Steven P. Sanderson II, MPH

Examples

ts_random_walk(
.mean = 6,
.sd = 1,
.num_walks = 25,
.periods = 180,
.initial_value = 6
)

#> # A tibble: 4,500 x 4
#>      run     x     y        cum_y
#>    <dbl> <dbl> <dbl>        <dbl>
#>  1     1     1  5.05         36.3
#>  2     1     2  5.92        252. 
#>  3     1     3  6.19       1808. 
#>  4     1     4  7.25      14914. 
#>  5     1     5  5.45      96214. 
#>  6     1     6  4.29     508938. 
#>  7     1     7  5.30    3205806. 
#>  8     1     8  7.90   28524708. 
#>  9     1     9  7.11  231457139. 
#> 10     1    10  6.85 1816235320. 
#> # ... with 4,490 more rows