Augment Function Velocity

Takes a numeric vector and will return the velocity of that vector.

Usage

ts_velocity_augment(.data, .value, .names = "auto")

Arguments

.data

The data being passed that will be augmented by the function.

.value

This is passed rlang::enquo() to capture the vectors you want to augment.

.names

The default is "auto"

Value

A augmented tibble

Details

Takes a numeric vector and will return the velocity of that vector. The velocity of a time series is computed by taking the first difference, so $$x_t - x_t1$$

This function is intended to be used on its own in order to add columns to a tibble.

See also

Author

Steven P. Sanderson II, MPH

Examples

suppressPackageStartupMessages(library(dplyr))

len_out    = 10
by_unit    = "month"
start_date = as.Date("2021-01-01")

data_tbl <- tibble(
  date_col = seq.Date(from = start_date, length.out = len_out, by = by_unit),
  a    = rnorm(len_out),
  b    = runif(len_out)
)

ts_velocity_augment(data_tbl, b)
#> # A tibble: 10 × 4
#>    date_col        a       b velocity_b
#>    <date>      <dbl>   <dbl>      <dbl>
#>  1 2021-01-01  1.97  0.666      NA     
#>  2 2021-02-01  0.284 0.0704     -0.596 
#>  3 2021-03-01  0.784 0.152       0.0820
#>  4 2021-04-01  2.34  0.0447     -0.108 
#>  5 2021-05-01  2.37  0.00778    -0.0369
#>  6 2021-06-01 -0.325 0.983       0.975 
#>  7 2021-07-01 -1.35  0.245      -0.738 
#>  8 2021-08-01 -0.191 0.131      -0.114 
#>  9 2021-09-01 -0.488 0.480       0.349 
#> 10 2021-10-01 -0.808 0.520       0.0394