Model Method Extraction Helper

This takes in a model fit and returns the method of the fit object.

Usage

model_extraction_helper(.fit_object)

Arguments

.fit_object

A time-series fitted model

Value

A model description

Details

Currently supports forecasting model of one of the following from the forecast package:

Arima

auto.arima

ets

nnetar

workflow fitted models.

See also

Author

Steven P. Sanderson II, MPH

Examples

# NOT RUN
if (FALSE) {
suppressPackageStartupMessages(library(forecast))

# Create a model
fit_arima  <- auto.arima(AirPassengers)

model_extraction_helper(fit_arima)
}