Compare data over time periods

Given a tibble/data.frame, you can get date from two different but comparative date ranges. Lets say you want to compare visits in one year to visits from 2 years before without also seeing the previous 1 year. You can do that with this function.

ts_compare_data(.data, .date_col, .start_date, .end_date, .periods_back)

Arguments

.data
The date.frame/tibble that holds the data

.date_col
The column with the date value

.start_date
The start of the period you want to analyze

.end_date
The end of the period you want to analyze

.periods_back
How long ago do you want to compare data too. Time units are collapsed using lubridate::floor_date(). The value can be:

second

minute

hour

day

week

month

bimonth

quarter

season

halfyear

year

Arbitrary unique English abbreviations as in the lubridate::period() constructor are allowed.

Value

A tibble.

Details

Uses the timetk::filter_by_time() function in order to filter the date column.

Uses the timetk::subtract_time() function to subtract time from the start date.

Author

Steven P. Sanderson II, MPH

Examples

suppressPackageStartupMessages(library(healthyR.data))
suppressPackageStartupMessages(library(dplyr))
suppressPackageStartupMessages(library(timetk))
ts_compare_data(
  .data           = healthyR_data
  , .date_col     = visit_start_date_time
  , .start_date   = "2019-01-01"
  , .end_date     = "2019-12-31"
  , .periods_back = "2 years"
  ) %>%
  select(visit_start_date_time) %>%
  summarise_by_time(
    .date_var = visit_start_date_time
    , .by     = "year"
    , visits  = n()
  )

#> # A tibble: 2 x 2
#>   visit_start_date_time visits
#>   <dttm>                 <int>
#> 1 2017-01-01 00:00:00    18873
#> 2 2019-01-01 00:00:00    17457

ts_compare_data(
 .data = healthyR_data
 , .date_col     = visit_end_date_time
 , .start_date   = "2019-01-01"
 , .end_date     = "2019-12-31"
 , .periods_back = "2 years"
)

#> # A tibble: 36,335 x 17
#>    mrn      visit_id   visit_start_date_ti~ visit_end_date_time total_charge_am~
#>    <chr>    <chr>      <dttm>               <dttm>                         <dbl>
#>  1 12437877 8607556387 2016-12-31 14:36:00  2017-01-01 00:00:00           13279.
#>  2 54652951 8715051466 2016-12-31 13:25:00  2017-01-01 00:00:00           12596.
#>  3 79070190 8324482206 2016-12-30 10:14:00  2017-01-01 00:00:00           34243.
#>  4 36365280 8952156137 2016-12-31 18:37:00  2017-01-01 00:00:00            9825.
#>  5 26710950 8838409607 2017-01-01 18:01:00  2017-01-01 00:00:00           18094.
#>  6 31297793 1399983322 2016-12-28 08:43:00  2017-01-01 10:11:00           52360.
#>  7 92856112 1341515963 2016-12-27 09:44:00  2017-01-01 10:45:00           94073.
#>  8 75781185 1691222926 2016-12-28 17:04:00  2017-01-01 11:57:00           51282.
#>  9 36570101 1859045312 2016-12-21 16:45:00  2017-01-01 12:39:00           72660.
#> 10 40236197 1471200535 2016-12-27 10:59:00  2017-01-01 13:41:00           59725.
#> # ... with 36,325 more rows, and 12 more variables: total_amount_due <dbl>,
#> #   total_adjustment_amount <dbl>, payer_grouping <chr>,
#> #   total_payment_amount <dbl>, ip_op_flag <chr>, service_line <chr>,
#> #   length_of_stay <dbl>, expected_length_of_stay <lgl>,
#> #   length_of_stay_threshold <lgl>, los_outlier_flag <dbl>, readmit_flag <dbl>,
#> #   readmit_expectation <lgl>