Time Series Splits Plot

Sometimes we want to see the training and testing data in a plot. This is a simple wrapper around a couple of functions from the timetk package.

ts_splits_plot(.splits_obj, .date_col, .value_col)

Arguments

.splits_obj

The predefined splits object.

.date_col

The date column for the time series.

.value_col

The value column of the time series.

Value

A time series cv plan plot

Details

You should already have a splits object defined. This function takes in three parameters, the splits object, a date column and the value column.

See also

Author

Steven P. Sanderson II, MPH

Examples

suppressPackageStartupMessages(library(modeltime))
suppressPackageStartupMessages(library(timetk))
suppressPackageStartupMessages(library(dplyr))
suppressPackageStartupMessages(library(healthyR.data))

data <- healthyR_data %>%
   filter(ip_op_flag == "I") %>%
   select(visit_end_date_time) %>%
   rename(date_col = visit_end_date_time) %>%
   summarise_by_time(
        .date_var = date_col
        , .by     = "month"
        , value   = n()
    ) %>%
   filter_by_time(
       .date_var     = date_col
       , .start_date = "2012"
       , .end_date   = "2019"
   )

splits <- time_series_split(
    data
    , date_col
    , assess = 12
    , skip = 3
    , cumulative = TRUE
)

ts_splits_plot(
    .splits_obj = splits,
    .date_col   = date_col,
    .value_col  = value
)